+++ src/app/common/business-models/property.business-model.ts
@@ -1,0 +1,10 @@
+import {AddressBusinessModel} from "./address.business-model";
+
+export interface PropertyBusinessModel {
+  id: string;
+  area: string;
+  roomsNumber: string;
+  description: string;
+  type: string;
+  address: AddressBusinessModel;
+}