+++ src/app/common/components/home-card/home-card.component.ts
@@ -2,10 +2,10 @@ import {Component, Input} from '@angular/core';
 
 @Component({
   selector: 'app-card',
-  templateUrl: './card.component.html',
-  styleUrls: ['./card.component.scss']
+  templateUrl: './home-card.component.html',
+  styleUrls: ['./home-card.component.scss']
 })
-export class CardComponent {
+export class HomeCardComponent {
   @Input() public title: string | undefined;
   @Input() public img: string | undefined;
   @Input() public alt: string | undefined;