+++ src/app/features/property/property-form/property-form.component.spec.ts
@@ -1,18 +1,18 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { CardComponent } from './card.component';
+import { PropertyFormComponent } from './property-form.component';
 
-describe('CardComponent', () => {
-  let component: CardComponent;
-  let fixture: ComponentFixture<CardComponent>;
+describe('PropertyFormComponent', () => {
+  let component: PropertyFormComponent;
+  let fixture: ComponentFixture<PropertyFormComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ CardComponent ]
+      declarations: [ PropertyFormComponent ]
     })
     .compileComponents();
 
-    fixture = TestBed.createComponent(CardComponent);
+    fixture = TestBed.createComponent(PropertyFormComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });